StyleCop, documentation fixes.
