+++ Managed.Adb.Extensions.Tests/Managed.Adb.Extensions.Tests.csproj
@@ -64,6 +64,7 @@
     <Compile Include="DummyDevice.cs" />
     <Compile Include="FileEntryTests.cs" />
     <Compile Include="FileSystemTests.cs" />
+    <Compile Include="LinuxPathTests.cs" />
     <Compile Include="ListingServiceReceiverTests.cs" />
     <Compile Include="ListingServiceTests.cs" />
     <Compile Include="PackageManagerTests.cs" />
