+++ Managed.Adb.Extensions/Exceptions/PackageInstallationException.cs
