+++ Managed.Adb.Extensions/FilePermissions.cs
