+++ Managed.Adb.Extensions/IO/LinuxPath.cs
