+++ Managed.Adb.Extensions/Receivers/InstallReceiver.cs
