+++ Managed.Adb.Extensions/ThrowIf.cs
