+++ Managed.Adb.Tests/SyncServiceTests.cs
@@ -42,7 +42,7 @@ namespace Managed.Adb.Tests
                 });
 
             Assert.IsNotNull(value);
-            Assert.AreEqual(SyncService.FileMode.Regular, value.FileMode & SyncService.FileMode.TypeMask);
+            Assert.AreEqual(UnixFileMode.Regular, value.FileMode & UnixFileMode.TypeMask);
             Assert.AreEqual(597, value.Size);
             Assert.AreEqual(ManagedAdbExtenstions.Epoch.ToLocalTime(), value.Time);
         }
