+++ Managed.Adb/ISyncService.cs
@@ -1,16 +1,12 @@
-﻿//-----------------------------------------------------------------------
-// <copyright file="ISyncService.cs" company="Quamotion">
-//     Copyright (c) 2015 Quamotion. All rights reserved.
+﻿// <copyright file="ISyncService.cs" company="The Android Open Source Project, Ryan Conrad, Quamotion">
+// Copyright (c) The Android Open Source Project, Ryan Conrad, Quamotion. All rights reserved.
 // </copyright>
-//-----------------------------------------------------------------------
 
 namespace Managed.Adb
 {
     using System;
     using System.Collections.Generic;
     using System.IO;
-    using System.Linq;
-    using System.Text;
     using System.Threading;
 
     /// <summary>
@@ -18,18 +14,22 @@ namespace Managed.Adb
     /// </summary>
     public interface ISyncService : IDisposable
     {
+        /// <include file='.\ISyncService.xml' path='/SyncService/IsOpen/*'/>
+        bool IsOpen { get; }
+
+        /// <include file='.\ISyncService.xml' path='/SyncService/Push/*'/>
         void Push(Stream stream, string remotePath, int permissions, IProgress<int> progress, CancellationToken cancellationToken);
 
-        /// <include file='.\ISyncService.xml' path='/SyncService/PullFile2/*'/>
-        void Pull(string remoteFilepath, Stream stream, IProgress<int> progress, CancellationToken cancellationToken);
+        /// <include file='.\ISyncService.xml' path='/SyncService/Pull/*'/>
+        void Pull(string remotePath, Stream stream, IProgress<int> progress, CancellationToken cancellationToken);
 
-        /// <include file='.\ISyncService.xml' path='/SyncService/Close/*'/>
-        void Close();
+        /// <include file='.\ISyncService.xml' path='/SyncService/Stat/*'/>
+        FileStatistics Stat(string remotePath);
+
+        /// <include file='.\ISyncService.xml' path='/SyncService/GetDirectoryListing/*'/>
+        IEnumerable<FileStatistics> GetDirectoryListing(string remotePath);
 
         /// <include file='.\ISyncService.xml' path='/SyncService/Open/*'/>
         void Open();
-
-        /// <include file='.\ISyncService.xml' path='/SyncService/IsOpen/*'/>
-        bool IsOpen { get; }
     }
 }
