+++ Managed.Adb/Managed.Adb.csproj
@@ -119,11 +119,11 @@
     <Compile Include="DeviceDataEventArgs.cs" />
     <Compile Include="DeviceState.cs" />
     <Compile Include="Exceptions\CommandAbortingException.cs" />
-    <Compile Include="Exceptions\PackageInstallationException.cs" />
+    <Compile Include="Exceptions\PermissionDeniedException.cs" />
     <Compile Include="Exceptions\ShellCommandUnresponsiveException.cs" />
     <Compile Include="Exceptions\UnknownOptionException.cs" />
     <Compile Include="Extensions\BytesHelper.cs" />
-    <Compile Include="FilePermissions.cs" />
+    <Compile Include="Extensions\StringHelper.cs" />
     <Compile Include="FileStatistics.cs" />
     <Compile Include="ForwardData.cs" />
     <Compile Include="IAdbClient.cs" />
@@ -134,42 +134,29 @@
     <Compile Include="ISyncService.cs" />
     <Compile Include="Logs\ILogListener.cs" />
     <Compile Include="Logs\LogEntry.cs" />
-    <Compile Include="Extensions\ThrowIf.cs">
-      <SubType>Code</SubType>
-    </Compile>
     <Compile Include="RawImage.cs" />
     <Compile Include="Receivers\ConsoleOutputReceiver.cs" />
     <Compile Include="DdmPreferences.cs" />
     <Compile Include="DeviceMonitor.cs" />
     <Compile Include="Exceptions\AdbException.cs" />
     <Compile Include="Exceptions\DeviceNotFoundException.cs" />
-    <Compile Include="Exceptions\PermissionDeniedException.cs" />
     <Compile Include="Extensions\ArrayHelper.cs" />
     <Compile Include="Extensions\DateTimeHelper.cs" />
-    <Compile Include="ErrorCodeHelper.cs" />
-    <Compile Include="Extensions\FileInfoHelper.cs" />
-    <Compile Include="Extensions\StringHelper.cs" />
-    <Compile Include="FilePermission.cs" />
     <Compile Include="ILogOutput.cs" />
-    <Compile Include="Receivers\InstallReceiver.cs" />
-    <Compile Include="IO\LinuxPath.cs" />
     <Compile Include="Receivers\IShellOutputReceiver.cs" />
-    <Compile Include="ISyncProgressMonitor.cs" />
-    <Compile Include="Receivers\LinkResolverReceiver.cs" />
     <Compile Include="Log.cs" />
     <Compile Include="LogLevel.cs" />
     <Compile Include="Logs\LogReceiver.cs" />
     <Compile Include="Receivers\MultilineReceiver.cs" />
     <Compile Include="Receivers\NullOutputReceiver.cs" />
-    <Compile Include="NullSyncProgressMonitor.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SocketExtensions.cs" />
     <Compile Include="StartServerResult.cs" />
     <Compile Include="SyncCommand.cs" />
     <Compile Include="SyncCommandConverter.cs" />
-    <Compile Include="SyncResult.cs" />
     <Compile Include="SyncService.cs" />
     <Compile Include="TransportType.cs" />
+    <Compile Include="UnixFileMode.cs" />
   </ItemGroup>
   <ItemGroup>
     <BootstrapperPackage Include="Microsoft.Net.Client.3.5">
@@ -207,6 +194,7 @@
     <Analyzer Include="..\packages\StyleCop.Analyzers.1.0.0-beta015\analyzers\dotnet\cs\StyleCop.Analyzers.CodeFixes.dll" />
     <Analyzer Include="..\packages\StyleCop.Analyzers.1.0.0-beta015\analyzers\dotnet\cs\StyleCop.Analyzers.dll" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(SolutionDir)\.nuget\NuGet.targets" Condition="Exists('$(SolutionDir)\.nuget\NuGet.targets')" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
