+++ Managed.Adb/Receivers/IShellOutputReceiver.cs
@@ -15,20 +15,6 @@ namespace Managed.Adb
     public interface IShellOutputReceiver
     {
         /// <summary>
-        /// Adds the output.
-        /// </summary>
-        /// <param name="data">The data.</param>
-        /// <param name="offset">The offset.</param>
-        /// <param name="length">The length.</param>
-        void AddOutput(byte[] data, int offset, int length);
-
-        /// <summary>
-        /// Flushes the output.
-        /// </summary>
-        /// <remarks>This should always be called at the end of the "process" in order to indicate that the data is ready to be processed further if needed.</remarks>
-        void Flush();
-
-        /// <summary>
         /// Gets a value indicating whether this instance is cancelled.
         /// </summary>
         /// <value>
@@ -47,5 +33,19 @@ namespace Managed.Adb
         /// will detect common error messages and throw an exception.
         /// </remarks>
         bool ParsesErrors { get; }
+
+        /// <summary>
+        /// Adds the output.
+        /// </summary>
+        /// <param name="data">The data.</param>
+        /// <param name="offset">The offset.</param>
+        /// <param name="length">The length.</param>
+        void AddOutput(byte[] data, int offset, int length);
+
+        /// <summary>
+        /// Flushes the output.
+        /// </summary>
+        /// <remarks>This should always be called at the end of the "process" in order to indicate that the data is ready to be processed further if needed.</remarks>
+        void Flush();
     }
 }
