+++ Managed.Adb/SyncCommandConverter.cs
@@ -8,36 +8,64 @@ namespace Managed.Adb
     using System.Collections.Generic;
     using System.Linq;
 
+    /// <summary>
+    /// Converts <see cref="SyncCommand"/> values to their binary representation
+    /// and vice versa.
+    /// </summary>
     public static class SyncCommandConverter
     {
-        private static readonly Dictionary<SyncCommand, string> values = new Dictionary<SyncCommand, string>();
+        /// <summary>
+        /// Maps the <see cref="SyncCommand"/> values to their string representations.
+        /// </summary>
+        private static readonly Dictionary<SyncCommand, string> Values = new Dictionary<SyncCommand, string>();
 
+        /// <summary>
+        /// Initializes static members of the <see cref="SyncCommandConverter"/> class.
+        /// </summary>
         static SyncCommandConverter()
         {
-            values.Add(SyncCommand.DATA, "DATA");
-            values.Add(SyncCommand.DENT, "DENT");
-            values.Add(SyncCommand.DONE, "DONE");
-            values.Add(SyncCommand.FAIL, "FAIL");
-            values.Add(SyncCommand.LIST, "LIST");
-            values.Add(SyncCommand.OKAY, "OKAY");
-            values.Add(SyncCommand.RECV, "RECV");
-            values.Add(SyncCommand.SEND, "SEND");
-            values.Add(SyncCommand.STAT, "STAT");
+            Values.Add(SyncCommand.DATA, "DATA");
+            Values.Add(SyncCommand.DENT, "DENT");
+            Values.Add(SyncCommand.DONE, "DONE");
+            Values.Add(SyncCommand.FAIL, "FAIL");
+            Values.Add(SyncCommand.LIST, "LIST");
+            Values.Add(SyncCommand.OKAY, "OKAY");
+            Values.Add(SyncCommand.RECV, "RECV");
+            Values.Add(SyncCommand.SEND, "SEND");
+            Values.Add(SyncCommand.STAT, "STAT");
         }
 
+        /// <summary>
+        /// Gets the byte array that represents the <see cref="SyncCommand"/>.
+        /// </summary>
+        /// <param name="command">
+        /// The <see cref="SyncCommand"/> to convert.
+        /// </param>
+        /// <returns>
+        /// A byte array that represents the <see cref="SyncCommand"/>.
+        /// </returns>
         public static byte[] GetBytes(SyncCommand command)
         {
-            if (!values.ContainsKey(command))
+            if (!Values.ContainsKey(command))
             {
                 throw new ArgumentOutOfRangeException(nameof(command), $"{command} is not a valid sync command");
             }
 
-            string commandText = values[command];
+            string commandText = Values[command];
             byte[] commandBytes = AdbClient.Encoding.GetBytes(commandText);
 
             return commandBytes;
         }
 
+        /// <summary>
+        /// Determines which <see cref="SyncCommand"/> is represented by this byte array.
+        /// </summary>
+        /// <param name="value">
+        /// A byte array that represents a <see cref="SyncCommand"/>.
+        /// </param>
+        /// <returns>
+        /// The corresponding <see cref="SyncCommand"/>.
+        /// </returns>
         public static SyncCommand GetCommand(byte[] value)
         {
             if (value == null)
@@ -52,7 +80,7 @@ namespace Managed.Adb
 
             string commandText = AdbClient.Encoding.GetString(value);
 
-            var key = values.Where(d => string.Equals(commandText, commandText, StringComparison.OrdinalIgnoreCase)).Select(d => new SyncCommand?(d.Key)).SingleOrDefault();
+            var key = Values.Where(d => string.Equals(commandText, commandText, StringComparison.OrdinalIgnoreCase)).Select(d => new SyncCommand?(d.Key)).SingleOrDefault();
 
             if (key == null)
             {
